added null pointer exception throws

+++ src/main/java/com/leon/controllers/MainController.java
@@ -31,6 +31,18 @@ public class MainController
     @RequestMapping(value = "/configuration", method=GET)
     public String getConfiguration(@RequestParam String owner, @RequestParam String key)
     {
+        if(owner == null)
+        {
+            logger.error("The owner request param cannot be null.");
+            throw new NullPointerException("owner");
+        }
+
+        if(key == null)
+        {
+            logger.error("The key request param cannot be null.");
+            throw new NullPointerException("key");
+        }
+
         logger.info("Received request for configuration value with owner: " + owner + ", and key: " + key);
         return configurationService.getConfigurationValue(owner, key);
     }
@@ -45,6 +57,12 @@ public class MainController
     @RequestMapping(value = "/configuration", method=POST)
     public void saveConfiguration(@RequestBody Configuration configuration)
     {
+        if(configuration == null)
+        {
+            logger.error("The configuration request body cannot be null");
+            throw new NullPointerException("configuration");
+        }
+
         logger.info("Received request to save a configuration: " + configuration);
         configurationService.saveConfiguration(configuration);
     }
@@ -52,6 +70,18 @@ public class MainController
     @RequestMapping(value = "/configuration", method=DELETE)
     public void deleteConfiguration(@RequestParam String owner, @RequestParam String key)
     {
+        if(owner == null)
+        {
+            logger.error("The owner request param cannot be null.");
+            throw new NullPointerException("owner");
+        }
+
+        if(key == null)
+        {
+            logger.error("The key request param cannot be null.");
+            throw new NullPointerException("key");
+        }
+
         logger.info("Received request to delete configuration value with owner: " + owner + ", and key: " + key);
         configurationService.deleteConfiguration(owner, key);
     }
